Fix several clang-tidy warnings, mostly related to integer conversions.
Most of the diff fixes the `width` parameter in `mem_{read,write}_callback`
to be an `int64_t` (Sail C API's `mach_int`) instead of `uint64_t`.

There were some missing casts for conversions in the Softfloat API.
Existing C-style casts were converted into `static_cast<>`.

Statically allocated constant version strings were converted into
C-style strings to fix a warning that the C++ string constructors
can raise exceptions during static allocation (induced by their use
to initialize `std::string_view`).  These are now accessed via
functions instead of as constants.

Other miscellaneous warnings were similarly fixed.

+++ c_emulator/riscv_softfloat.cpp
@@ -358,7 +358,7 @@ bv5_bv16 softfloat_i32tof16(uint64_t rm, uint64_t v) {
   SOFTFLOAT_PRELUDE(rm);
 
   float16_t res;
-  res = i32_to_f16((int32_t)v);
+  res = i32_to_f16(static_cast<int32_t>(v));
 
   return {softfloat_exceptionFlags, res.v};
 }
@@ -367,7 +367,7 @@ bv5_bv16 softfloat_ui32tof16(uint64_t rm, uint64_t v) {
   SOFTFLOAT_PRELUDE(rm);
 
   float16_t res;
-  res = ui32_to_f16((uint32_t)v);
+  res = ui32_to_f16(static_cast<uint32_t>(v));
 
   return {softfloat_exceptionFlags, res.v};
 }
@@ -376,7 +376,7 @@ bv5_bv16 softfloat_i64tof16(uint64_t rm, uint64_t v) {
   SOFTFLOAT_PRELUDE(rm);
 
   float16_t res;
-  res = i64_to_f16(v);
+  res = i64_to_f16(static_cast<int64_t>(v));
 
   return {softfloat_exceptionFlags, res.v};
 }
@@ -394,7 +394,7 @@ bv5_bv32 softfloat_i32tof32(uint64_t rm, uint64_t v) {
   SOFTFLOAT_PRELUDE(rm);
 
   float32_t res;
-  res = i32_to_f32((int32_t)v);
+  res = i32_to_f32(static_cast<int32_t>(v));
 
   return {softfloat_exceptionFlags, res.v};
 }
@@ -403,7 +403,7 @@ bv5_bv32 softfloat_ui32tof32(uint64_t rm, uint64_t v) {
   SOFTFLOAT_PRELUDE(rm);
 
   float32_t res;
-  res = ui32_to_f32((uint32_t)v);
+  res = ui32_to_f32(static_cast<uint32_t>(v));
 
   return {softfloat_exceptionFlags, res.v};
 }
@@ -412,7 +412,7 @@ bv5_bv32 softfloat_i64tof32(uint64_t rm, uint64_t v) {
   SOFTFLOAT_PRELUDE(rm);
 
   float32_t res;
-  res = i64_to_f32(v);
+  res = i64_to_f32(static_cast<int64_t>(v));
 
   return {softfloat_exceptionFlags, res.v};
 }
@@ -430,7 +430,7 @@ bv5_bv64 softfloat_i32tof64(uint64_t rm, uint64_t v) {
   SOFTFLOAT_PRELUDE(rm);
 
   float64_t res;
-  res = i32_to_f64((int32_t)v);
+  res = i32_to_f64(static_cast<int32_t>(v));
 
   return {softfloat_exceptionFlags, res.v};
 }
@@ -439,7 +439,7 @@ bv5_bv64 softfloat_ui32tof64(uint64_t rm, uint64_t v) {
   SOFTFLOAT_PRELUDE(rm);
 
   float64_t res;
-  res = ui32_to_f64((uint32_t)v);
+  res = ui32_to_f64(static_cast<uint32_t>(v));
 
   return {softfloat_exceptionFlags, res.v};
 }
@@ -448,7 +448,7 @@ bv5_bv64 softfloat_i64tof64(uint64_t rm, uint64_t v) {
   SOFTFLOAT_PRELUDE(rm);
 
   float64_t res;
-  res = i64_to_f64(v);
+  res = i64_to_f64(static_cast<int64_t>(v));
 
   return {softfloat_exceptionFlags, res.v};
 }
